Split unit tests into smaller files
